#174 make persons search case insensitive (+refactor and improve tests)
